requirements.txt
asgiref==3.6.0
astroid==2.15.2
bcrypt==4.0.1
blinker==1.6
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dill==0.3.6
dnspython==2.3.0
dominate==2.7.0
email-validator==1.3.1
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-MySQL==1.5.2
flask-mysql-connector==1.1.0
Flask-MySQLdb==1.0.1
flask-oidc==1.4.0
Flask-WTF==1.1.1
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nodeenv==1.7.0
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.0.0
platformdirs==3.2.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pylint==2.17.2
PyMySQL==1.0.3
pyparsing==3.0.9
pyright==1.1.302
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tomlkit==0.11.7
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
visitor==0.1.3
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
